requirements.txt
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Maple==0.5.5
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Pillow==5.4.1
requests==2.21.0
